:zap: Fixed tests for older elixir versions
Turns out adding a test for pow `**` was not a brilliant idea
considering it was added in 1.13 lol.

Also had to implement a flag to test configurations on different states.
It's pretty damn hard to modify config at compile time because... well
it probably compiles the files on parallel so race conditions.

Hadn't noticed until I decided to do a force recompile on test heh